% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{hello}
\usage{
hello(n = 5, name = "world")
}
\arguments{
\item{n}{Specifies the number of times to repeat the message. Default value is 5.}

\item{name}{Who to say hello to. Default value is "world".}
}
\value{
Prints a friendly message.
}
\description{
This function prints a friendly message to the user
}
\examples{
hello()

}

