% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{plot_scatter}
\alias{plot_scatter}
\title{plot_scatter}
\usage{
plot_scatter(n = 100)
}
\arguments{
\item{n}{The number of points to plot, default is 100.}
}
\value{
A scatter plot of the data.
}
\description{
Prints a scatter plot of randomly generated data
}
\examples{
plot_scatter(n = 200)
}

